Adds student call-on feature to seating chart
Implements a feature to track which students have been called on in a class seating chart.

Introduces the ability to mark students as "called on" and highlights them visually.
Provides functionality to reset the call-on status for all students and to pick a random student that hasn't been called on yet.